Split Pg into per-category traits under skunk.sharp.pg.functions
`PgFunction.scala` was growing toward 500+ lines with aggregates, math,
string, time, null, and subquery functions all in one object. Split:

  skunk.sharp.pg.functions/
    Shared.scala        -- shared types (StrLike, Lift, SumOf, AvgOf)
                           and the `sameTypeFn` / `stringPreserveFn` /
                           `doubleFn` / `stringToIntFn` shape helpers.
    PgNumeric.scala     -- trait: math (abs, ceil, floor, sqrt, power, ...)
    PgString.scala      -- trait: lower, upper, trim, replace, substring, ...
    PgTime.scala        -- trait: now, currentTimestamp, currentDate, ...
    PgAggregate.scala   -- trait: count*, sum, avg, min, max, stringAgg, ...
    PgNull.scala        -- trait: coalesce, nullif
    PgSubquery.scala    -- trait: exists, notExists

  skunk.sharp.Pg        -- object Pg extends PgNumeric with PgString
                           with PgTime with PgAggregate with PgNull
                           with PgSubquery

`PgFunction.scala` is now ~65 lines — just the `PgFunction` / `PgOperator`
arity helpers.

The mixin shape opens a clean extension point: third-party modules can
ship their own `JsonbFns` / `LtreeFns` / ... traits and either extend
`object Pg` indirectly via an `object MyPg extends PgNumeric with ...
with JsonbFns`, or just expose their own namespace (`object Jsonb`).

No behaviour / test change — 145 core tests green.
